refactor: Ladder와 LadderRow가 Players에 직접 의존하지 않도록 수정
