+++ src/main/java/domain/ladder/LadderRow.java
@@ -6,8 +6,6 @@ import static domain.ladder.DirectionalPoint.STRAIGHT;
 import static domain.ladder.DirectionalPoint.findDirectionalPoint;
 
 import domain.booleanGenerator.BooleanGenerator;
-import domain.player.Player;
-import domain.player.Players;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -21,14 +19,6 @@ public class LadderRow {
             boolean isTheLastPoint = currentIndex == rowSize - 1;
             DirectionalPoint ladderPoint = determineNextPoint(booleanGenerator, getLastPoint(), isTheLastPoint);
             ladderPoints.add(ladderPoint);
-        }
-    }
-
-    // TODO: Ladder - Players 결합도 낮추기
-    public void playRow(Players players) {
-        for (int i = 0; i < ladderPoints.size(); i++) {
-            Player player = players.findPlayerByIndex(i);
-            player.move(ladderPoints.get(player.getPosition()).getDirection());
         }
     }
