+++ src/test/java/domain/ladder/LadderRowTest.java
@@ -4,9 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import domain.booleanGenerator.BooleanGenerator;
-import domain.player.Name;
-import domain.player.Player;
-import domain.player.Players;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import support.TrueGenerator;
@@ -41,31 +38,6 @@ public class LadderRowTest {
                 () -> assertThat(ladderPoint.get(1)).isEqualTo(DirectionalPoint.LEFT),
                 () -> assertThat(ladderPoint.get(2)).isEqualTo(DirectionalPoint.RIGHT),
                 () -> assertThat(ladderPoint.get(3)).isEqualTo(DirectionalPoint.LEFT)
-        );
-    }
-
-    @Test
-    void 각_플레이어는_가로대를_만나면_이동한다() {
-        // given
-        Players players = new Players(List.of(
-                new Player(new Name("name1"), 0),
-                new Player(new Name("name2"), 1),
-                new Player(new Name("name3"), 2),
-                new Player(new Name("name4"), 3)
-        ));
-
-        // |=====|     |=====|
-        LadderRow ladderRow = new LadderRow(trueGenerator, players.getPlayerCount());
-
-        // when
-        ladderRow.playRow(players);
-
-        // then
-        assertAll(
-                () -> assertThat(players.findPlayerByIndex(0).getPosition()).isEqualTo(1),
-                () -> assertThat(players.findPlayerByIndex(1).getPosition()).isEqualTo(0),
-                () -> assertThat(players.findPlayerByIndex(2).getPosition()).isEqualTo(3),
-                () -> assertThat(players.findPlayerByIndex(3).getPosition()).isEqualTo(2)
         );
     }
 }
