+++ src/test/java/domain/ladder/LadderTest.java
@@ -1,13 +1,8 @@
 package domain.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
 
 import domain.booleanGenerator.BooleanGenerator;
-import domain.player.Name;
-import domain.player.Player;
-import domain.player.Players;
-import java.util.List;
 import org.junit.jupiter.api.Test;
 import support.TrueGenerator;
 
@@ -25,35 +20,5 @@ public class LadderTest {
 
         // then
         assertThat(ladder.getLadderRows().size()).isEqualTo(height.getValue());
-    }
-
-    @Test
-    void 각_플레이어는_사다리를_타고_이동한다() {
-        // given
-        Height height = new Height(3);
-        Players players = new Players(List.of(
-                new Player(new Name("name1"), 0),
-                new Player(new Name("name2"), 1),
-                new Player(new Name("name3"), 2),
-                new Player(new Name("name4"), 3)
-        ));
-
-        /*
-         *  |=====|     |=====|
-         *  |=====|     |=====|
-         *  |=====|     |=====|
-         */
-        Ladder ladder = new Ladder(trueGenerator, height, players.getPlayerCount());
-
-        // when
-        ladder.play(players);
-
-        // then
-        assertAll(
-                () -> assertThat(players.findPlayerByIndex(0).getPosition()).isEqualTo(1),
-                () -> assertThat(players.findPlayerByIndex(1).getPosition()).isEqualTo(0),
-                () -> assertThat(players.findPlayerByIndex(2).getPosition()).isEqualTo(3),
-                () -> assertThat(players.findPlayerByIndex(3).getPosition()).isEqualTo(2)
-        );
     }
 }
